requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
streamlit>=1.38.0
requests>=2.31.0
